+++ tests/metrics/test_default.py
@@ -89,3 +89,19 @@ class TestDefaultAccuracy(unittest.TestCase):
         mock.process_final.assert_called_once_with(4, 5, 6)
         self.assertEqual(mock.eval.call_count, 1)
         self.assertEqual(mock.train.call_count, 1)
+
+    @patch('torchbearer.metrics.default.CategoricalAccuracy')
+    def test_reset_after_eval(self, cat_acc):
+        metric = DefaultAccuracy()
+        self.assertTrue(cat_acc.call_count == 1)
+        cat_acc.reset_mock()
+        metric.eval()
+
+        state = {torchbearer.CRITERION: F.cross_entropy}
+        mock = Mock()
+        mock.eval = Mock()
+        torchbearer.metrics.default.__loss_map__[F.cross_entropy.__name__] = lambda: mock
+
+        metric.reset(state)
+
+        self.assertTrue(mock.eval.call_count == 1)